refactor: separated server and app so that during testing while the main server is running there is no port usage conflict. Test cases can now run while server / port is being used.
