+++ back-end/tests/integration/authRoutes.test.js
@@ -1,10 +1,11 @@
 const request = require("supertest");
-const app = require("../../server");
+const app = require("../../app");
 
 describe("Auth API Tests", () => {
 	let testEmail = `test${Date.now()}@test.com`;
 	let token = "";
 
+	// Register a new user
 	it("should register a new user", async () => {
 		const res = await request(app).post("/auth/register").send({ email: testEmail, password: "password123" });
 
@@ -12,6 +13,7 @@ describe("Auth API Tests", () => {
 		expect(res.body.message).toEqual("User registered successfully");
 	});
 
+	// Prevent duplicate email registrations
 	it("should prevent duplicate email registrations", async () => {
 		const res = await request(app).post("/auth/register").send({ email: testEmail, password: "password123" });
 
@@ -19,20 +21,27 @@ describe("Auth API Tests", () => {
 		expect(res.body.message).toEqual("User already exists");
 	});
 
+	// Log in the user
 	it("should log in the user and return a token", async () => {
 		const res = await request(app).post("/auth/login").send({ email: testEmail, password: "password123" });
 
 		expect(res.statusCode).toEqual(200);
 		expect(res.body.accessToken).toBeDefined();
+
 		token = res.body.accessToken;
 	});
 
-	it("should get the user's information with valid token", async () => {
+	// Get user information with a valid token
+	it("should get the user's information with a valid token", async () => {
 		const res = await request(app).get("/auth/user").set("Authorization", `Bearer ${token}`);
+
+		console.log(res.body.message);
 
 		expect(res.statusCode).toEqual(200);
 		expect(res.body.email).toEqual(testEmail);
 	});
+
+	// Invalid token test
 	it("should return 401 for invalid token", async () => {
 		const res = await request(app).get("/auth/user").set("Authorization", `Bearer invalidtoken`);
